+++ polo_client/lib/polo_web_client.dart
@@ -1,6 +1,5 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
-library polo_client;
+/// PoloWeb Client
+
+library polo_web_client;
 
 export 'src/polo_web_client_helper.dart';