+++ polo_client/lib/src/polo_web_client_base.dart
@@ -35,7 +35,7 @@ class PoloWebClient {
   }
 
   /// Closes the connection to the `PoloServer`
-  Future<void> close() async {
+  void close() async {
     return _webSocket.close();
   }