+++ polo_client_ts/src/types.ts
@@ -1,0 +1,5 @@
+// deno-lint-ignore-file no-explicit-any
+type VoidCallback = () => void;
+type DataCallback = (data: any) => void;
+
+export type { DataCallback, VoidCallback };